files fixed for parcel
