+++ .cache/484a576d211c6e85c8e1a53cb2c21ba7.json
@@ -1,0 +1,1 @@
+{"id":"game.html","dependencies":[{"name":"./keyboarder.js","dynamic":true,"parent":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/game.html","resolved":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/keyboarder.js"},{"name":"./game.js","dynamic":true,"parent":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/game.html","resolved":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/game.js"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n    <meta charset=\"UTF-8\">\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n\n    <title>Hank's Game</title>\n    <style>\n        body {\n            background-color: #000010;\n        }\n        .game {\n        width: 550px;\n        height: 550px;\n        margin: 20px auto;\n        background-color: black;\n        margin-bottom: 40px;\n        }\n        h1 {\n            text-align: center;\n            font-family: Arial, Helvetica, sans-serif;\n            color: white;\n        }\n        .writing {\n            color: white;\n            text-align: center;\n            font-family: Arial, Helvetica, sans-serif;\n        }\n    </style>\n</head>\n    <body>\n        <h1 class\"header\"=\"\">Not Space Invaders</h1>\n        <div class=\"game\">\n        <canvas id=\"game-canvas\" width=\"550\" height=\"550\"></canvas>\n        </div>   \n        <p class=\"writing\">Shit</p>\n        <script src=\"/6e5dde5839415fd552850ec28f28d4a3.js\"></script>\n        <script src=\"/ca84566ab0744f4f6ee9e07c49beb541.js\"></script>\n    </body>\n</html>"},"hash":"79a8d87c3a0c88b5269bfc00513dc757","cacheData":{}}
