+++ .cache/6440d55efa94d69499bd324011e8bef0.json
@@ -1,0 +1,1 @@
+{"id":"src/constants.js","dependencies":[],"generated":{"js":"'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nvar canvas = exports.canvas = document.getElementById('game-canvas');\nvar screenSize = exports.screenSize = { x: canvas.width, y: canvas.height\n};","map":{"mappings":[{"generated":{"line":6,"column":0},"source":"src/constants.js","original":{"line":1,"column":7}},{"name":"canvas","generated":{"line":6,"column":4},"source":"src/constants.js","original":{"line":1,"column":13}},{"name":"document","generated":{"line":6,"column":30},"source":"src/constants.js","original":{"line":1,"column":22}},{"name":"getElementById","generated":{"line":6,"column":39},"source":"src/constants.js","original":{"line":1,"column":31}},{"generated":{"line":6,"column":53},"source":"src/constants.js","original":{"line":1,"column":22}},{"generated":{"line":6,"column":54},"source":"src/constants.js","original":{"line":1,"column":46}},{"generated":{"line":6,"column":67},"source":"src/constants.js","original":{"line":1,"column":22}},{"generated":{"line":6,"column":68},"source":"src/constants.js","original":{"line":1,"column":7}},{"generated":{"line":7,"column":0},"source":"src/constants.js","original":{"line":2,"column":7}},{"name":"screenSize","generated":{"line":7,"column":4},"source":"src/constants.js","original":{"line":2,"column":13}},{"generated":{"line":7,"column":38},"source":"src/constants.js","original":{"line":2,"column":26}},{"name":"x","generated":{"line":7,"column":40},"source":"src/constants.js","original":{"line":2,"column":28}},{"name":"canvas","generated":{"line":7,"column":43},"source":"src/constants.js","original":{"line":2,"column":31}},{"name":"width","generated":{"line":7,"column":50},"source":"src/constants.js","original":{"line":2,"column":38}},{"generated":{"line":7,"column":55},"source":"src/constants.js","original":{"line":2,"column":26}},{"name":"y","generated":{"line":7,"column":57},"source":"src/constants.js","original":{"line":2,"column":45}},{"name":"canvas","generated":{"line":7,"column":60},"source":"src/constants.js","original":{"line":2,"column":48}},{"name":"height","generated":{"line":7,"column":67},"source":"src/constants.js","original":{"line":2,"column":55}},{"generated":{"line":8,"column":0},"source":"src/constants.js","original":{"line":2,"column":26}},{"generated":{"line":8,"column":1},"source":"src/constants.js","original":{"line":2,"column":7}}],"sources":{"src/constants.js":"export const canvas = document.getElementById('game-canvas')\nexport const screenSize = { x: canvas.width, y: canvas.height\n}\n"},"lineCount":null}},"hash":"a8b6adee3010ac725bb6a932e1669447","cacheData":{"env":{}}}
