+++ .cache/ceb3571b16243146626478633afde959.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"./src/game.js","dynamic":true,"parent":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/index.html","resolved":"/Users/hankmitsch/Documents/MomentumLabs/project-w3-fe-game-hankmitsch/src/game.js"}],"generated":{"html":"<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n    <meta charset=\"UTF-8\">\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n    <meta http-equiv=\"X-UA-Compatible\" content=\"ie=edge\">\n\n    <title>Hank's Game</title>\n    <style>\n        body {\n            background-color: #000010;\n        }\n        .game {\n        width: 550px;\n        height: 550px;\n        margin: 20px auto;\n        background-color: black;\n        margin-bottom: 40px;\n        }\n        h1 {\n            text-align: center;\n            font-family: Arial, Helvetica, sans-serif;\n            color: white;\n        }\n        .writing {\n            color: white;\n            text-align: center;\n            font-family: Arial, Helvetica, sans-serif;\n        }\n    </style>\n</head>\n    <body>\n        <h1 class\"header\"=\"\">Not Space Invaders</h1>\n        <div class=\"game\">\n        <canvas id=\"game-canvas\" width=\"550\" height=\"550\"></canvas>\n        </div>   \n        <p class=\"writing\">Shit</p>\n        <script src=\"/6f84b53db2c1f2479688e8e95ea2b669.js\"></script>\n    </body>\n</html>"},"hash":"089fb32aca77adf84da8997d1cde49a3","cacheData":{}}
