+++ src/constants.js
@@ -1,0 +1,3 @@
+export const canvas = document.getElementById('game-canvas')
+export const screenSize = { x: canvas.width, y: canvas.height
+}
