+++ src/keyboarder.js
@@ -25,3 +25,5 @@ class Keyboarder {
 }
 
 Keyboarder.KEYS = { LEFT: 37, RIGHT: 39, UP: 38, DOWN: 40, S: 83 }
+
+export default Keyboarder
